Add files via upload
Program that calculates Grade point average of any scaling system( some have negative grades like USA system and others don't like the Canadian system i.e. (university of Toronto), and also accounts for different scales ( of 4 or of 5).
A file is attached with errors that i made and how i changed them, the mentioned file along with the whole file I guess are for employment purposes, they are intended as an experience based proof of proficiency in python but feel free too contribute and learn along from it with me.